requirements/main.txt
python-dateutil>=2.8.2
regex>=2023.5.5
PyYAML>=6.0
jsonata-python>=0.6.0
